requirements.txt
python-dotenv
streamlit>=1.35
plotly
mp-api
pymatgen
numpy
pandas
